requirements.txt
fastapi
uvicorn
python-dotenv
requests
firebase-admin
